Fix a bunch of leaks and use std.os.system.poll()
Apparently std.io.poll() is an "abstraction" over std.os.system.poll(),
which has a very unintuitive name. The helpful people over at zigcord
let me know that the "syscall" is the proper way to use this, so let's
do that.

This also fixes a bunch of borked leaked allocations, because RAII is
evil and the programmer is not an idiot (apparently).