Ensure correct thread sequencing and handle loop events first
As we're using threads to substitute for an event loop, we don't want
*actual* concurrency here, this commit makes sure the threads run in
lockstep.
Also handle the loop queue events before the deferred actions, in case
stdin has data but we've opted to exit to retry; this fixes a wayward
read() blocking the repl when the very first read event causes the
editor to quit.